requirements.txt
asgiref==3.8.1
boto3==1.35.77
botocore==1.35.77
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.2.6
packaging==24.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.2
uritemplate==4.1.1
urllib3>=1.25.4,<1.27
